Pass in large dfs, accomedate single & multiple states
Passing in large dataframes rather than a path speeds up the whole process. For the NERC calculations I was passing in a list of states and using .isin() to filter the facility dataframe. This didn't work with single string values, so I'm converting single 2-letter state codes into a list.