Change input parameters and comment out change since 2005
As with the other function in this file, I changed some input parameters from paths to dataframes. This reduces the number of times that a file needs to be loaded from disk. It also lets me change column names before passing through if necessary.

Because some NERC regions don't have data for 2005, I commented out the calculation of change in index since 2005. Maybe this should be included as an optional parameter in the future.