Clean up some cells, plot high CO2/capita states on separate plot
The goal is to create an inset plot with the highest CO2/capita states. I can probably figure out how to do this in matplotlib

Also added data on net flow of electricity in/out of states. This helps to show how much electricity was exported in states like WY.